Dia 3

+++ Dia 3/main.py
@@ -1,3 +1,3 @@
-num = float(input("Digite um número: "))
+num = int(input("Digite um número: "))
 
 print(f"O numero digitado {num} é {'par' if num % 2 == 0 else 'ímpar'}")
